Update from https://github.com/Anime-Manga/references/commit/c5ee5162f6da9a76c05c0259229700cfbbbb5b26

+++ src/references/Cesxhin.AnimeManga.Domain/DTO/ChapterDTO.cs
@@ -6,7 +6,7 @@ namespace Cesxhin.AnimeManga.Domain.DTO
     {
         public string ID { get; set; }
         public string NameManga { get; set; }
-        public int CurrentVolume { get; set; }
+        public float CurrentVolume { get; set; }
         public float CurrentChapter { get; set; }
         public int NumberMaxImage { get; set; }
         public string UrlPage { get; set; }
